requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
discord.py==2.3.2
dnspython==2.4.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
pip==23.2.1
pymongo==4.5.0
PyYAML==6.0.1
requests==2.31.0
setuptools==68.0.0
urllib3==2.0.4
wheel==0.38.4
yarl==1.9.2
